// src/app/main/model/produto.ts
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Produto { 
    id?: number;
    nome?: string | null;
    descricao?: string | null;
    preco?: number;
    sldAtual?: number;
    categoria?: string | null;
}